+++ docker/Dockerfile
@@ -16,10 +16,10 @@ ENV UV_LINK_MODE=copy \
   UV_PYTHON=python3.12 \
   UV_PROJECT_ENVIRONMENT=/opt/venv \
   PATH="/opt/venv/bin:$PATH"
-WORKDIR /opt/stactools-ephemeral
+WORKDIR /opt/stactools-osc-hazard
 COPY --from=base /opt/conda /opt/conda
 COPY pyproject.toml ./
-COPY src/stactools/ephemeral/__init__.py src/stactools/ephemeral/
+COPY src/stactools/osc_hazard/__init__.py src/stactools/osc_hazard/
 RUN apt-get -y -q update \
   && apt-get -y -q install build-essential \
   && rm -rf /var/lib/apt/lists/
@@ -34,15 +34,15 @@ ENV UV_LINK_MODE=copy \
   UV_PYTHON=python3.12 \
   UV_PROJECT_ENVIRONMENT=/opt/venv \
   PATH="/opt/venv/bin:$PATH"
-WORKDIR /opt/stactools-ephemeral
+WORKDIR /opt/stactools-osc-hazard
 COPY --from=base /opt/conda /opt/conda
 COPY --from=dependencies /opt/venv /opt/venv
 COPY pyproject.toml ./
 COPY src ./src
 RUN uv sync --no-dev --no-editable
 WORKDIR /
-RUN rm -rf /opt/stactools-ephemeral
-CMD [ "stac", "ephemeralcmd" ]
+RUN rm -rf /opt/stactools-osc-hazard
+CMD [ "stac", "oschazard" ]
 
 
 FROM dependencies as dev-dependencies
@@ -53,7 +53,7 @@ ENV UV_LINK_MODE=copy \
   UV_PYTHON=python3.12 \
   UV_PROJECT_ENVIRONMENT=/opt/venv \
   PATH="/opt/venv/bin:$PATH"
-WORKDIR /opt/stactools-ephemeral
+WORKDIR /opt/stactools-osc-hazard
 COPY --from=dependencies /opt/venv /opt/venv
 COPY pyproject.toml .
 RUN uv sync \
@@ -63,8 +63,8 @@ RUN uv sync \
 FROM dev-dependencies as dev-builder
 
 ENV PATH="/opt/venv/bin:$PATH"
-WORKDIR /opt/stactools-ephemeral
+WORKDIR /opt/stactools-osc-hazard
 COPY --from=base /opt/conda /opt/conda
 COPY --from=dependencies /opt/venv /opt/venv
 COPY . .
-CMD [ "stac", "ephemeralcmd" ]
+CMD [ "stac", "oschazard" ]
