+++ docker/lint
@@ -15,7 +15,7 @@ Run code formatters in a docker container with all prerequisites installed.
 
 if [ "${BASH_SOURCE[0]}" = "${0}" ]; then
     docker run --rm -it \
-        -v "$(pwd)":/opt/stactools-ephemeral \
+        -v "$(pwd)":/opt/stactools-osc-hazard \
         --entrypoint scripts/lint \
-        stactools-packages/ephemeral:dev
+        stactools-packages/osc-hazard:dev
 fi
