+++ environment.yml
@@ -1,4 +1,4 @@
-name: stactools-ephemeral
+name: stactools-osc-hazard
 channels:
   - conda-forge
   - defaults
