+++ pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "stactools-ephemeral"
+name = "stactools-osc-hazard"
 version = "0.1.0"
 description = "PROVIDE DESCRIPTION HERE"
 readme = "README.md"
@@ -34,8 +34,8 @@ dev = [
 docs = ["pystac~=1.10", "ipykernel~=6.29", "jinja2~=3.1"]
 
 [project.urls]
-Github = "https://github.com/stactools-packages/ephemeral"
-Issues = "https://github.com/stactools-packages/ephemeral/issues"
+Github = "https://github.com/stactools-packages/osc-hazard"
+Issues = "https://github.com/stactools-packages/osc-hazard/issues"
 
 [build-system]
 requires = ["setuptools", "wheel"]
