+++ scripts/update-examples
@@ -5,13 +5,13 @@ from pathlib import Path
 
 from pystac import CatalogType
 
-import stactools.ephemeral.stac
+import stactools.osc_hazard.stac
 
 root = Path(__file__).parents[1]
 examples = root / "examples"
 
-collection = stactools.ephemeral.stac.create_collection()
-item = stactools.ephemeral.stac.create_item(str(root / "tests" / "data" / "asset.tif"))
+collection = stactools.osc_hazard.stac.create_collection()
+item = stactools.osc_hazard.stac.create_item(str(root / "tests" / "data" / "asset.tif"))
 collection.add_item(item)
 collection.update_extent_from_items()
 collection.normalize_hrefs(str(examples))
