+++ src/stactools/osc_hazard/__init__.py
@@ -1,6 +1,6 @@
 import stactools.core
 from stactools.cli.registry import Registry
-from stactools.ephemeral.stac import create_collection, create_item
+from stactools.osc_hazard.stac import create_collection, create_item
 
 __all__ = ["create_collection", "create_item"]
 
@@ -8,6 +8,6 @@ stactools.core.use_fsspec()
 
 
 def register_plugin(registry: Registry) -> None:
-    from stactools.ephemeral import commands
+    from stactools.osc_hazard import commands
 
-    registry.register_subcommand(commands.create_ephemeralcmd_command)
+    registry.register_subcommand(commands.create_oschazard_command)
