+++ src/stactools/osc_hazard/stac.py
