+++ tests/test_commands.py
@@ -4,11 +4,11 @@ from click import Group
 from click.testing import CliRunner
 from pystac import Collection, Item
 
-from stactools.ephemeral.commands import create_ephemeralcmd_command
+from stactools.osc_hazard.commands import create_oschazard_command
 
 from . import test_data
 
-command = create_ephemeralcmd_command(Group())
+command = create_oschazard_command(Group())
 
 
 def test_create_collection(tmp_path: Path) -> None:
