+++ tests/test_stac.py
@@ -1,4 +1,4 @@
-from stactools.ephemeral import stac
+from stactools.osc_hazard import stac
 
 from . import test_data
